requirements.txt
Django==1.8.5
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.1.1
static==0.4
wsgiref==0.1.2
djangorestframework==3.2.4
psycopg2==2.6.1
wheel==0.24.0
